+++ src/org/corfield/new/impl.clj
@@ -210,24 +210,6 @@
       [repo (and path root) (or path root repo) tag]))
   )
 
-(def ^:private clojure-test-runner
-  {:test-runner/coordinate "io.github.cognitect-labs/test-runner {:git/tag \"v0.5.1\" :git/sha \"dfb30dd\"}"
-   :test-runner/main       "cognitect.test-runner"
-   :test-runner/exec-fn    "cognitect.test-runner.api/test"
-   :test-runner/namespace  "clojure.test"
-   :test-runner/deftest    "deftest"
-   :test-runner/is         "is"
-   :test-runner/testing    "testing"})
-
-(def ^:private lazytest-runner
-  {:test-runner/coordinate "io.github.noahtheduke/lazytest {:mvn/version \"1.8.0\"}"
-   :test-runner/main       "lazytest.main"
-   :test-runner/exec-fn    "lazytest.main/run-impl"
-   :test-runner/namespace  "lazytest.core"
-   :test-runner/deftest    "defdescribe"
-   :test-runner/is         "expect"
-   :test-runner/testing    "it"})
-
 (defn preprocess-options
   "Given the raw options hash map, preprocess, parse, and
   validate certain values, and derive defaults for others."
@@ -273,24 +255,8 @@
             :target-dir target-dir
             :user       username
             :version    "0.1.0-SNAPSHOT"}
-           (when (= :bb (:build opts))
-             {::babashka true})
-           ;; make test runner swappable:
-           (if (= :lazytest (:test-runner opts))
-             lazytest-runner
-             clojure-test-runner)
            ;; remove options we cleaned up:
            (dissoc opts :template :target-dir :name))))
-
-(defn maybe-add-bb
-  "If the user has asked for a Babashka template, update the template.edn
-  to include `bb.edn` and use the simpler `build.clj` file."
-  [edn data]
-  (if (::babashka data)
-    (-> edn
-        (assoc-in  [:transform 0 0] "build-bb")
-        (update-in [:transform 0 2] assoc "bb.tmpl" "bb.edn"))
-    edn))
 
 (defn get-multi-option
   "Given a hash map of options, return the value for the
@@ -321,12 +287,12 @@
                        ;; :data-fn result is additive:
                        (merge opts ((requiring-resolve data-fn) opts)))
                      basic-opts
-                     (get-multi-option basic-opts :data-fn))
+                     (get-multi-option basic-edn :data-fn))
         edn  (reduce (fn [edn template-fn]
                        ;; :template-fn result is replacement:
                        ((requiring-resolve template-fn) edn opts))
                      basic-edn
-                     (get-multi-option opts :template-fn))]
+                     (get-multi-option basic-edn :template-fn))]
     ;; this allows any defaults from the template to
     ;; be part of the data used for substitution:
     [(merge {:description (str "FIXME: my new"